Adds the test script used to verify the improved generator
The compareOutput.sh script takes in the number of universities to
generate (defaulting to 10 if not set) and generates the data using both
the original and the improved generator.  For the improved generator it
generates data in both single-threaded and multi-threaded mode and using
the --consolidate option.  Data is generated in all the common formats
(OWL, N-Triples and Turtle).

The generated data is then compared against the generated data from the
original generator to verify that the outputs are identical.