Improve how Full consolidation does IO
Previous implementation of Full consolidation could result in poor CPU
utilisation and unecessary blocking on IO particularly when compression
was used.

This commit rewrites the Full consolidation IO handling into a separate
threaded service that handles all the writes in the background allowing
generator threads to continue working on generating more data while we
wait for the IO to complete.  IO is also now batched up as far as
possible which can further improve performance.