BackgroundWriterService error handling tweaks
If the BackgroundWriterService thread dies or is terminated it could
leave generator threads blocked waiting to put things into the write
queue if the write queue is full.  To avoid this when exiting the writer
thread clear the queue which allows outstanding put() calls to succeed
and any subsequent write submissions will produce an error.