Fix a possible OOM when using Full consolidation at large scale
Due to how we do the buffering the Full consolidation mode could exhaust
the heap because buffers we were finished with had references to them
kept unecessarily.  This change ensures we throw out the buffers when we
are done with them.

It also improves error handling so that the code will fail faster if a
data generator fails.